修复队列长度不能大于255bug;

+++ Repeater/Repeater_Project/class/AudioQueue/AudioQueue.h
@@ -44,7 +44,7 @@ typedef RepeaterAudioPacket*	P_RepeaterAudioBuffer;
 #pragma pack(push, 1)
 typedef struct{
 	char		data[320];
-	uint8_t		len;
+	int			len;
 
 }fifoqueue_t;
